Share the live-loopback harness across the web python suites
The register-server/client preamble, the result-capture sink pair, and
the port-triggered client request were repeated across test_runtime and
test_endpoints (twice); they now live in _loopback_harness.py.  This is
the deduplicable slice of the SonarCloud new-code duplication finding.
